requirements.txt
# Data Science & ML Model
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter
ipykernel
xgboost

# FastAPI Backend & Server
fastapi
uvicorn
pydantic
python-multipart

# Database & Migration
SQLAlchemy
python-dotenv
psycopg2-binary
